[editor] Renaming Components Prior to Packaging
