+++ python/src/aiconfig/editor/client/src/index.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import Editor from "./Editor";
+import LocalEditor from "./LocalEditor";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 root.render(
   <React.StrictMode>
-    <Editor />
+    <LocalEditor />
   </React.StrictMode>
 );
